refactor: move alerts into their own folder
