+++ packages/curve-ui-kit/src/features/manage-soft-liquidation/ui/alerts/AlertClosePosition.tsx
